move progress from lab to core
