+++ .changeset/empty-guests-tell.md
@@ -1,0 +1,5 @@
+---
+"@salt-ds/lab": minor
+---
+
+Removed `CircularProgress` and `LinearProgress` from lab and promoted to core.
