+++ packages/core/src/__tests__/__e2e__/progress/LinearProgress.cy.tsx
