+++ packages/core/stories/progress/progress.qa.stories.tsx
@@ -1,10 +1,10 @@
 import { Meta, StoryFn } from "@storybook/react";
 
-import { CircularProgress, LinearProgress } from "@salt-ds/lab";
+import { CircularProgress, LinearProgress } from "@salt-ds/core";
 import { QAContainer, QAContainerProps } from "docs/components";
 
 export default {
-  title: "Lab/Progress/Progress QA",
+  title: "Core/Progress/Progress QA",
   component: CircularProgress,
 } as Meta<typeof CircularProgress | typeof LinearProgress>;
 
@@ -13,12 +13,14 @@ export const ExamplesGrid: StoryFn<QAContainerProps> = (props) => {
   return (
     <QAContainer cols={2} {...props}>
       <LinearProgress
+        aria-label="Download"
         className={className}
         value={38}
         style={{ padding: "50px" }}
       />
       <CircularProgress aria-label="Download" value={38} />
       <LinearProgress
+        aria-label="Download"
         className={className}
         value={38}
         style={{ padding: "50px" }}
