+++ packages/core/stories/progress/progress.stories.css
