+++ packages/lab/src/index.ts
@@ -57,7 +57,6 @@ export * from "./overlay";
 export * from "./parent-child-item";
 export * from "./parent-child-layout";
 export * from "./portal";
-export * from "./progress";
 export * from "./query-input";
 export * from "./responsive";
 export * from "./search-input";
