+++ site/docs/components/progress/index.mdx
@@ -7,13 +7,13 @@ data:
 
   # Fill in the info from the content template's "Metadata" table below.
   # To omit optional items, comment them out with #
-  sourceCodeUrl: "https://github.com/jpmorganchase/salt-ds/tree/main/packages/lab/src/progress"
+  sourceCodeUrl: "https://github.com/jpmorganchase/salt-ds/tree/main/packages/core/src/progress"
   package:
-    name: "@salt-ds/lab"
+    name: "@salt-ds/core"
   alsoKnownAs: [Progress bar, Progress indicator]
   relatedComponents: []
   # stickerSheet: "https://figma.com/..."
-  status: Lab component
+  status: Core component
 # Leave this as is
 layout: DetailComponent
 ---
