+++ site/docs/components/progress/usage.mdx
@@ -24,20 +24,20 @@ data:
 
 ### Import
 
-To import the `CircularProgress` and `LinearProgress` components from the lab Salt package, use:
+To import the `CircularProgress` and `LinearProgress` components from the core Salt package, use:
 
 ```js
-import { CircularProgress } from "@salt-ds/lab";
+import { CircularProgress } from "@salt-ds/core";
 // or
-import { LinearProgress } from "@salt-ds/lab";
+import { LinearProgress } from "@salt-ds/core";
 ```
 
 ### Props
 
 #### `CircularProgress`
 
-<PropsTable packageName="lab" componentName="CircularProgress" />
+<PropsTable packageName="core" componentName="CircularProgress" />
 
 #### `LinearProgress`
 
-<PropsTable packageName="lab" componentName="LinearProgress" />
+<PropsTable packageName="core" componentName="LinearProgress" />
