+++ site/src/examples/progress/Circular.tsx
@@ -1,5 +1,5 @@
 import { ReactElement } from "react";
-import { CircularProgress } from "@salt-ds/lab";
+import { CircularProgress } from "@salt-ds/core";
 
 export const Circular = (): ReactElement => (
   <CircularProgress aria-label="Download" value={38} />
