+++ site/src/examples/progress/Linear.tsx
@@ -1,5 +1,5 @@
 import { ReactElement } from "react";
-import { LinearProgress } from "@salt-ds/lab";
+import { LinearProgress } from "@salt-ds/core";
 
 export const Linear = (): ReactElement => (
   <LinearProgress aria-label="Download" value={38} />
