+++ site/src/examples/progress/WithMaxVal.tsx
@@ -1,11 +1,13 @@
 import { ReactElement, useState } from "react";
-import { CircularProgress, LinearProgress } from "@salt-ds/lab";
 import {
   FlexItem,
   FlexLayout,
   FlowLayout,
+  H3,
   RadioButton,
   RadioButtonGroup,
+  CircularProgress,
+  LinearProgress,
 } from "@salt-ds/core";
 
 export const WithMaxVal = (): ReactElement => {
@@ -13,7 +15,7 @@ export const WithMaxVal = (): ReactElement => {
 
   return (
     <FlexLayout direction="column" style={{ height: "100%" }}>
-      <h3 style={{ textAlign: "center" }}> max = 500, value = 250</h3>
+      <H3 style={{ textAlign: "center" }}> max = 500, value = 250</H3>
       <FlexItem>
         <FlowLayout justify="center" className="controls" gap={1}>
           <RadioButtonGroup direction="horizontal" defaultChecked>
